+++ modules/extensions-xml-support/src/test/java/org/mule/test/functional/AbstractXmlExtensionMuleArtifactFunctionalTestCase.java
@@ -9,8 +9,8 @@ package org.mule.test.functional;
 import static org.mule.runtime.api.dsl.DslResolvingContext.getDefault;
 import static org.mule.runtime.core.api.util.boot.ExtensionLoaderUtils.getLoaderById;
 import static org.mule.runtime.extension.api.ExtensionConstants.XML_SDK_LOADER_ID;
-import static org.mule.runtime.extension.internal.loader.XmlExtensionModelLoader.RESOURCES_PATHS;
-import static org.mule.runtime.extension.internal.loader.XmlExtensionModelLoader.RESOURCE_DECLARATION;
+import static org.mule.runtime.extension.internal.feeder.XmlExtensionModelLoader.RESOURCES_PATHS;
+import static org.mule.runtime.extension.internal.feeder.XmlExtensionModelLoader.RESOURCE_DECLARATION;
 
 import static java.util.Optional.empty;
 
@@ -23,8 +23,8 @@ import org.mule.runtime.core.api.config.ConfigurationBuilder;
 import org.mule.runtime.core.api.config.builders.AbstractConfigurationBuilder;
 import org.mule.runtime.core.api.extension.ExtensionManager;
 import org.mule.runtime.extension.api.loader.ExtensionModelLoader;
-import org.mule.runtime.extension.internal.loader.XmlExtensionLoaderDelegate;
-import org.mule.runtime.extension.internal.loader.XmlExtensionModelLoader;
+import org.mule.runtime.extension.internal.feeder.XmlExtensionLoaderDelegate;
+import org.mule.runtime.extension.internal.feeder.XmlExtensionModelLoader;
 
 import java.util.HashMap;
 import java.util.HashSet;
